Module coco_keyprovider::enc_mods

Modules

crypto 🔒

kbs 🔒

Structs

AnnotationPacket

AnnotationPacket is what a encrypted image layer’s org.opencontainers.image.enc.keys.provider.attestation-agent annotation should contain when it is encrypted by CoCo’s encryption modules. Please refer to issue https://github.com/confidential-containers/attestation-agent/issues/113

InputParams 🔒

Constants

DEFAULT_KEY_REPO_PATH 🔒

When a KEK is randomly generated, a new kid will be generated with this prefix.

HARD_CODED_KEYID 🔒

KBS_RESOURCE_URL_PREFIX 🔒

Functions

enc_optsdata_gen_anno

The input params vector should only have one element. The format of the element is in the following format:

generate_key_parameters 🔒

This function will generate (key, iv, keyid) for given InputParams

normalize_path 🔒

Normalize the given keyid into (kbs addr, key path), s.t. converting kbs://... or ../.. to (<kbs-addr>, <repository>/<type>/<tag>).

parse_input_params 🔒